qa/870: cut systemctl some slack
When run with --check (as is the case from my check.callback), be more
lenient with systemctl and treat the state "activating" as the same as
"active", so don't report an error.

I've seen several cases where systemctl is "activating" and after
several seconds (4 or 5) is still waiting for the pmlogger lock as in
    pmsleep -w waiting for lock 0.1
is at the end of the systemctl process trace that starts with
pmlogger_check.

This is probably due to a timing collision between QA scripts and the
regular systemctl housekeeping for pmlogger, and until we have evidence
to the contrary should not cause a QA failure in the callback routine.