src/pmlogger/pmlogger_daily.sh: last part (I think) of $(date...) work
The previous commits had re-instated the behaviour of embedded shell
commands in the directory field of the control file, for example:
    /gpfs/data/root/pcp/$(date +%Y)/$(date +%m)/LOCALHOSTNAME/$(date +%Y)-$(date +%m)-$(date +%d)

What was not done was the work to make pmlogger_daily look in other
directories for archives to chew on, because they were hiding in
directories different from today's directory.

This commit fixes that.

There are a bunch of QA changes, mostly remade .out files as a result
of pmlogger_daily now using find(1) in places where ls(1) in the
current (today's) directory used to suffice ... this leads to some
different ordering in the sets of archives to be processed.

qa/1473 is new.