qa: bpf PMDA tests now _notrun if modules cannot be loaded
After a lot of patience and fruitless triage, I'm now whacking the bpf
PMDA QA so that if the BPF modules needed for a test cannot be loaded
the test will be _notrun.  The repeated failures of tests in this class
usually end up being some botch between the kernel, kernel modules,
libbpf or the toolchain ... in short nothing to do with the PCP code.
And the existing _notrun guards in the test cases are not sufficient
to detect these scenarios, so the tests fail.

Specifically the changes involve a new function _pmdabpf_tryload()
added to common.bpf which uses the (recently added) --tryload option
for pmdabpf to parse a config file and load the associated modules. If
this fails, _notrun() is called.

Then all the bpf PMDA test cases are modified to call
_pmdabpf_tryload() before trying to install the PMDA.