qa/1339: revert previous change, port 5670, ... are fine for *BSD
Not sure what the original analysis for the change to port 95670 was,
but the PMDA install failed with that tcp port on vm37 (OpenBSD 7.1),
and reverting the 5670 allowed the test to pass multiple times.

So there is some unexplained timing issue or port re-use issue here
which has not been fixed.