libpcp & pmlogrewrite: add PM_CTXFLAG_METADATA_ONLY flag for pmNewContext()
When processing compressed archives, opening an archive may start
decompression on all compressed parts (typically the .meta file and
the .0 volume) ... even before any read() is done.

For pmlogrewrite use from pmlogger_daily we need to ensure the archives
are consistent before any merging is attempted, so pmlogrewrite is
called with the -q flag to check metadata consistency.

Unfortunately if xz compression is in play (the default) liblzma
will start to uncompress the .0.xz file, even though pmlogrewrite is
not going to read that file if there are no changes to be made (the
typical case for recently created archives).  If the data volumes are
large this may create enormous memory pressure, CPU burn and blows
out the elapsed time for pmlogrewrite.

A new PM_CTXFLAG_METADATA_ONLY flag for pmNewContext() half-opens the
archive with *ONLY* the metadata file available ... this is sufficient
for pmlogrewrite -q on the fast-path, no changes path.

Simple tests suggest this change will reduce the running time for
pmlogrewrite -q by as much as a factor of 75 (and maybe more for very
large archives).