podman PMDA: do not use root_podman for pmns
root_<pmda> pmns files are reserved for the PMDAs that are
configured during the build and are unconditionally installed.

Optional PMDAs like this one should have a root and pmns files in
the PMDAs directory and any entry in the PMNS directory comes from
the Install script (not part of a package installation).