pmdadenki: remove unreachable code relating to sign (covscan)
Coverity static code checking noticed that there's a check for
negative power_now values in the denki PMDA - this cannot ever
happen, however, because an unsigned integer type is used when
reading and exporting this value.  As a result its safe to not
perform this additional check (current code is unreachable).