pmdasockets: allow sockets with duplicate sources to exist
An incorrect assumption was made in pmdasockets that the socket src
field is unique. This is not necessarily the case and when multiple
ss lines share a source only one will win (the final one). To solve
this we extend the instance names to include the destination.

A minor improvement to formatting in pcp-ss is also made, so that a
very long source field will not merge into the preceding column; we
always want to separate columns in case of parsing but it is also a
visual improvement too when such long socket source fields occur.

Resolves Red Hat BZ #2134413