libpcp: refactor pmDupContext()
There has been a TODO comment in the libpcp source code that
pmDupContext() is not really threadsafe (although the circumstances
to expose this are very unlikely).  However Coverity noticed, as per
CIDs 407600 and 407605.

So, I've refactored pmDupContext() to play nicely in the presence
of threads, but in the process exposed a bug in the original
implementation that has been there, I think, for TWENTY-NINE years.
Specifically, with the original implementation, after returning from
pmDupContext() the instance profile of the new context may not be
identical to the instance profile of the original context ... this
is plain wrong.

In trying to track this down (qa/004 and qa/005 were failing,
everything else was passing), I eventially hardened up
qa/src/context_test.c to check the instance profile of the dup'd
context and the original context, and with the unchanged libpcp,
kaboom.

When pmDupContext() was behaving correctly, it turns out that there
was a second latent bug in the QA test program that caused it to pass
with the bad libpcp and fail with the good libpcp.

This commit includes:
- the refactored pmDupContext()
- the extra integrity checking in qa/src/context_test.c
- bug fix for qa/src/context_test.c so it works correctly when
  pmDupContext() behaves properly.