qa/667 & qa/667.out: sample.datasize is sort of broken, filter problem away
sample.datasize is not really the datasize of the sample PMDA
because it is not initialized until the metric is fetched, and
it reports growth in the data segment since that point in time.

This test is exercising pcp2graphite, so tha value of sample.datasize
does not really matter.