pmOpenLog: refactor .prev handling
This routine is mostly used by PMDAs and pm* tools to create log files to
capture diagnostic information.

There were a number of ad hoc approaches to trying to save the previous
version of a the log file logname in logname.prev ... these were being done
in shell scripts run under various user id's and subject to some security
concerns.

The libpcp routine pmOpenLog now assumes responsibility for salvaging the
previous logname, if any, in logname.prev.

The shell scripts are not longer responsible for this activity.