rc scripts: need to ensure $PCP_RUN_DIR exists
In the scripts, while we're still root:root we need to ensure $PCP_RUN_DIR
exists, is a directory, is owned pcp:pcp and has mode 775.

This whole thing is messy because $PCP_RUN_DIR can be any of the following:
1. a persistent directory
2. a directory in a mounted tmpfs
3. a directory that is cleared out by reboot

In addition we may be using systemd (no wrappers for the non-pmcd ones) or
not.

These changes are really a defensive approach that are mostly for the
non-systemd platforms.