pmlogrewrite (libpcp, pmlogredact): fix problem when data volumes missing
Nathan discovered a problem with pmlogredact where for one input
archive the metadata in the output archive was truncated (mostly
missing).  Analysis showed that data volume 0 (the first one) for
the input archive was missing.

Since pmlogredact is a wrapper arounf pmlogrewrite, the real issue
was lurking in pmlogrewrite.

First, we need a libpcp API change for __pmLogCreate() (this is an
internal routine, so not part of the public pmapi.h APIs), so that
an extra argument can be passed to specify the initial data volume
number (it was hard-coded to 0, which is fine for everyone except
pmlogrewrite in the failure scenario described above).

Second, in pmlogrewrite itself the logic for processing the input
metadata needs a heuristic to suspend reading metadata to prevent the
metadata processing from getting ahead of the data (volume) processing
... this heuristic broke in the failure scenario described above and
the processing of the input metadata was suspended and never restarted.
The heuristic has been reworked in the case where one or more input
data volumes is found to be missing.