src/external/sort_r.h: refine FreeBSD change
Turns out that FreeBSD 14 changed the qsort_r() API to
match most of the rest of the world (Microsoft of course is
different to everyone else).

So need to re-instate the old "arse-backwards arguments" code
for earlier FreeBSD versions (was failing ... builds but pmcd
segv's when pmlogger starts [don't even think about it] ...
on vm06, FreeBSD 13.2).