pcp-reboot-init: new systemd service for $PCP_RC_DIR setup
a) consolidate all the $PCP_RC_DIR care and feeding (especially
   creating with appropriate mode and ownership after reboot)
   into a new pcp-reboot-init script (in $PCP_BINADM_DIR)
b) all the rc_wrapper scripts (which run as root) call
   pcp-reboot-init
c) for systemd systems a new pcp-reboot-init.service ensures
   the pcp-reboot-init scripts is run before any of the pmlogger,
   pmie or pmproxy services (which run as the user pcp)

Together with the tmpfiles.d change, this ensures $PCP_RC_DIR is
properly set up for systemd-based systems and those not using systemd,
and independent if which PCP "services" (pmcd, pmlogger, pmie and
pmproxy) might be enabled and active.

Also this setup has to happen after every system reboot, and after any
PCP software install or upgrade.

This script also provides a hook that can be used in the future for
any other housekeeping that's similar in semantics to the
$PCP_RC_DIR case and needs to be run as root.