qa/admin/pcp-daily: expand warnings filter
Want to re-focus on build warnings, so cull the things we can't
or won't fix, e.g.
- bogus warnings from old compilers
- non-PCP source, especially from the vendor dir and all of libpcp_qwt
- false matches for words like "warning"

For reference, in the QA Farm, the old filter reported 10,958 build
errors or warnings across 22 hosts.  After these filter changes, the
number of unique build errors or warnings is reduced to 198 ... still
too many, but within range of being attackable.