src/pcp-reboot-init/pcp-reboot-init.sh: stat(1) is not the same everywhere
Of course! For OpenBSD 7.3 (at least) the command line arguments for
stat(1) are not the same as for Linux, especially -t which means
something TOTALLY different.

Choose ls(1) instead and rely on user and group names, not numbers.

And there is no group "root" for BSDland, so qa/1478 needed some
rework.