qa/1441 & qa/common.check: fix race condition for pmie startup
qa/1441 was failing on vm11 (Debian 10.2) when the initial PID
for the newly started pmie was '' ... added guard and diagnostic
here that exposed root cause to be a race between pmie start up and
_wait_for_pmie() such that $PCP_RUN_DIR/pmie.pid had been created,
bit pmie had not yet written its PID into the file.

Fixed _wait_for_pmie() but left guard in qa/1441, to be sure,
to be sure.