qa/1384.out: "unsupported feature bits ..." messages have gone away
These messages from pmlogrewrite and pmlogdump vanished when the
PM_CTXFLAG_NO_FEATURE_CHECK flag was added to pmNewContext() on 14
Mar 2022 (!) ... failure of this QA test was masked by the default
valgrind strategy that chooses to run either the valgrind variant or
the non-valgrind variant but not both.

Unfortunately, the valgrind variant of this test only checks for
memory leaks, and the functional pass/fail is filtered away.

Forcing the non-valgrind version (1384) to be run exposed the "issue"
which is simply a remake for 1384.out.