src/pmdas/summary/mainloop.c: send state change 0 (no change) back to pmcd
When the PMDA-pmcd protocol was revamped to overload PMDA state
changes into the (previously unused) timestamp of a pmResult
that the PMDA sends to pmcd, we missed the summary PMDA which
has its own special mainloop to hide Seppo's secret agent mode
for pmie.

The result is that the summary PMDA was sending uninitialized
date in the byte of the timestamp that is supposed to hold
PMDA state change information.

In concert with a related bug in pmcd (see the next commit), this
meant the summary PMDA could cause badness to pmcd and even
pmcd clients ... in the failure that triggered this triage the
summary agent was able to cause the primary pmlogger to terminate,
c/o qa/525 ... but only sometimes on some hosts.