src/pmcd/src/dofetch.c: harden up ExtractState()
The first byte of the (historically unused) timestamp in a pmResult
sent from a PMDA to pmcd is reserved for encoding PMDA state changes
that pmcd needs to know about and/or tell pmcd's clients about.

pmcd was bravely assuming the PMDAs were all playing by the rules,
but the summary PMDA was sending rubbish in the state change byte
causing all manner of havoc.

Add filtering in ExtractState() to warn about bogus state change
values and cull any bits that a PMDA is not supposed to set, ever.