qa/766: deal with multiple interfaces from host(1)
On some platforms in CI, host(1) is returning multiple "has address"
lines, and so our $addr is set to multiple words,  e.g.
addr=10.1.0.5
127.0.1.1
which breaks the pmfind call later on.

So cull the 127.0.1.<anything> addrress, and if we still have more than
one ip addr (broken network config, or may be stripped interfaces?)
pick the first one and warn in .full