build: fix c-lint checks with bits/xxx includes
We never want to directly include bits/xxx include files,
so remove this cppcheck workaround.  Working backward from
there we need to guard against a cppcheck bug in the way
it expands glibs headers instead - and we can't do this in
a way that requires updating for every Fedora release, its
too obscure (caused wasted time for several folk).  So, in
the end, simplify the script a bunch and see how we go ...