src/libpcp/logmeta.c: fix SEGV with V3 archives
Problem was in __pmGetInDomArchive_ctx() where the test for big indoms
and optional local hashed access contained an early break that could
prevent the __pmLogUndeltaInDom() which meant we were left processing
a delta format indom which the code below assumed would never happen.

Based on ad hoc testing, I assert this fixes
    https://github.com/performancecopilot/pcp/issues/1904
but we don't have any QA test that tickles the precondition, and
pcp-atopsar only seems to fail on an active (or very large) zeroconf
type archive.