pmcheck plugins: overhead PMDA and pcp-zeroconf package
These two are not available for non-Linux systems, so for systems
like FreeBSD install an alternate plugin that can inform the use why
these components cannot be controlled with pmcheck.

This is a proof of concept, the same approach can be applied to
other "components" that are not unniversally available.