qa/common.bcc: update _try_compile()
On vm17 (openSUSE Leap 15.6) there appears to be yet another way
for the bcc PMDA to fail miserably ... this time an attempt to
compile and load (apparently) any module produces the message
"LLVM ERROR:" with not additional information!

Not suprisingly all the bcc PMDA tests were failing on this
platform.