src/pmlogcheck/pmlogcheck.c: make order of processing deterministic
The components of the archive used to be processed in directory
entry order, which is
(a) not human-friendly, and
(b) not QA-friendly.

Finally annoyed me to the point where I've added qsort() and
a knowledgeable compar() to sort files in this order:
- index
- meta
- data volumes is ascending volume number sequence

Allows qa/566 to be simplified and qa/566.out makes a lot more
sense as it no longer has output piped into sort(1) to make the
answers deterministic (pmlogcheck now deals with that).