libpcp: small derived metrics bug fix
This is a partial fix for
https://github.com/performancecopilot/pcp/issues/1921

If the operand has the semantics of PM_SEM_COUNTER and the
value goes backwards from one fetch to the next, then return
no values for rate().

There is QA to come, but this needs a final blessing from Oracle
before the reproducer archive can be included.