pmrep conf: fix and update pmrep proc metricsets
Use (more) correct rate() for total IO/s formula, also avoid reusing
the derived metric name already defined in the libpcp derived config.

Drop the IO % metric which was supposed to show how large a portion
the process is responsible for all of the IO. The formula did not
work on many occasions when per-process IO statistics were updated
in real-time but the totals (disk.all.*) only when the kernel really
wrote the data to disks.

In theory one might be able to calculate this by summing all the
process IO but the derived metric language does not allow doing a
calculation like delta(x) / sum(x) and this would also need some
pmrep changes to create an exception when to fetch all instance
values even if only reporting for few selected ones. Currently
this does not seem like worth the effort.