libpcp: derived metrics: add novalue() and defined(x) ? ... rework
Marko raised this (entirely reasonable) use case where we want to
return "no values" if metric "x" is undefined, else some expression
involving "x".

First step is a new novalue() function added to the derived metric
"grammar".

Second is special case handling of the construct:
    defined(x) ? <expr-involving-x> : novalue()
so that <expr-involving-x> is not evaluated if defined(x) is false.

And the dual for the construct:
    ! defined(x) ?  novalue() : <expr-involving-x>

This commit makes all of this working and a (new) qa/1508 to check
it out.

What's missing is some semantic checks to detect badness at "bind"
time, e.g. if neither <left-expr> nor <right-expr> are valid for a
    <guard> ?  <left-expr> : <right-expr>
construct.

And the man page needs some tweaking to remove the differences between
what I thought I'd implement and the actual implementation.