libpcp: derived metrics lazy binding for defined(x) ? ... : ...
Related to the previous novalue() extension, this commit extends
the ? ternary operator to allow lazy binding of metrics (name lookup
is not fatal) and conditional evaluation of only <left-expr> or
<right-expr>.

This allows
    defined(foo) ? foo : bar
to be valid both when foo exists in the PMNS and when foo does not
exist in the PMNS.  Previously this would fail at bind-time if foo
was not in the PMNS, which defeated the purpose of the expression,
and a consequence the construct was never used.

man page and qa/1508 changes to match.