man pages & more mandoc "lint" insanity
Those responsible for mandoc -T lint have so little clue about
troff that they think whitespace at the END OF A COMMENT is grounds
for breaking the PCP build.  [deep] Sigh.

And because Brian is not the Messiah, naughty-boy did not process
the symlinks below man/man_src and so this was missed for pcp-dstat.5.

The rest is some more man-spell fiddling about.