man-spell: audit-man-dict (new)
audit-man-dict checks the dictionary embedded in man-spell and finds
words in the dictionary that are not mentioned in any man page.

This list can be used to cull (mark as unused) those words in the
dictionary to speed up man-spell.  The first pass of this reduced
the man-spell dictionary from 2372 words to 1345 words.