libpcp: derived metrics: rework ? ... : ... lazy binding
There were some corner cases where the logic in check_expr()
was NQR ... fix this.

When __dmbind failed, there was some mangled logic on the
error paths that meant the expression tree was not always
dumped when __dmbind failed and with -Dderive,appl1 in play
... fix this.

And a bunch of QA remakes because the diagnostics for the
error test cases are now a little different sometimes.