libpcp et. al.: derived metric saga continues
This is a bit of a saga, but I've included chapter and verse for those with
an interest in software engineering, QA engineering, history, heroic Norse
legends or archaeology.

Starting with the derived disk.* metrics in src/derived/iostat.conf,
I wanted to refactor the expressions to use some of the newer derived
metrics features, in particular novalue() and meta=<metric> for novalue()
and mkconst().

The expressions in iostat.conf are complicated because they are trying
to defend against operand metrics not being available (as is the case on
older platforms, or non-Linux platforms) and divide by zero ... this leads
to nesting of ternary expressions of the form: <guard> ? <left> : <right>

First the (new) lazy binding meant that when <guard> can be evaluated at
bind time, trivially so for the common defined(x) case, then the "false"
branch of <left> or <right> is not evaluated.  This is necessary to allow
expresions of the form: defined(x) ? <expr-involving-x> : novalue() to
work independent of the availability of x.

Now add the special semantics when <left> is singular and <right> is
defined over an indom (or vice versa), which demands that if the singular
expression is chosen, then it provides a value for *all* the instances
in the indom from the *other* expression, e.g. <some-true-guard> ? 42 :
sample.bin => 9 values for <left>

But lazy binding means there may be *no* evaluation on the indom side of
the : and if the other expression is mkconst(... meta=x) and x has an indom,
were totally screwed.

The fix is to recognize this special condition and fetch the instance
domain to populate the mkconst() expression, so call pmGetInDom() ...
but we already have the ctxp->c_lock held at this point, so that's a nested
lock scenario and you're screwed again.

Never mind, the pmFoo_ctx() duals for pmFoo() were invented for just this
case ... so pmGetInDom_ctx() is the go ... wait, when the pmFoo_ctx()
refactoring was done almost 6 years ago pmGetInDom was omitted for no
good reason that I recall (git blame in internal.h will point the finger
... but that's a bit too cyclic so let's move on).

With out libpcp changes (internal.h, instance.c, derive_fetch.c, exports.in)
in place, back to the western front, where in the QA nomansland we find
there is no archive suitable for exercising all the dev.* derived metrics
from iostat.conf.  To get the complete set I need disk.dm.* available,
so I need a /dev/md/... device.  Off to vm03, only to find F38 has been
retired, so a couple of system upgrades later and I'm looking at F40.

I want 2 SCSI devices on different SCSI buses ... but virt-manager that
I use for VM admin on bozo refuses to do this.  No problem, shutdown the
vm03 virtual machine and edit the VM's XML directly, ... start vm03 again
... nothing's change ... repeat ... no change ... remember the XML is
really cached, so you need "$ virsh edit vm03" to really make the change
... start vm03 and the reboot hangs.  No amount of rescue magic helps,
because I've half created the RAID1 md device and then screwed with the I/O
architecture to the point where the kernel cannot recover the md device.
Blow vm03 away and do a clean install of F40 from scratch on vm03, ...
discover small error in packing manifest for F40.  Get Makepkgs working
again and recreate the md device and we're off to the races.

Create qa/archives/mk.disk-{a,b} to create qa/archives/disk-{a,b} archives
for qa/1509 (new).  First attempt *includes* the derived disk.* metrics
in the pmlogger configuration file ... this has an odd side-effect that
metrics like disk.dev.await end up in the archive, but with their PMID
rewritten from 511.0.22 to 511.2048.22.  Now when the qa test tries to
play with disk.dev.await from the disk-a archive, do we get the 511.2048.22
metric from the archive, or the 511.0.22 one from iostat.conf?

[insert 2 days of head pounding against brick wall]

I remember *why* the undocumented pmlogger weirdness is like this, but I
have no clue why we do not jump up and down with warnings when we appear
to have 2 different PMIDs for the same metric name in the PMNS.

Now qa/1509 can be used to pare away at the available metrics one operand
at a time to see what happens the associated derived metrics that involve
that operand ... and the result is not pretty; the refactored iostat.conf
expressions are incomplete or wrong, there are corner cases melt-downs in
both check_expr() and eval_expr() that these expressions are triggering.

Once that is all fixed, QA check -g derive passes [soft applause], but
then I discover that derived metrics are hiding in lots of other places
and -g derive is only running about half of the QA tests that could (and
did) fail.  Lots of cleanup in qa/group, and ... I think we're done.

But no, pmrep's :sar-d-dm config is now generating output (instead of
failing) for qa/1428 ... headlamp on for a forensic deep dive, yep
seems ok.