configure.ac: don't build Perl modules for some 32-bit platforms
We're moving to _TIME_BITS=64 ... but Perl is lagging behind, e.g.
on Debian 12.5 (vm21), and Perl has been built with a 32-bit time_t
and we can't make our Perl modules work with this and PCP libraries
that are using 64-bit time_t.

So Perl loses when this happens and we will stop building the Perl
bits in PCP.