src/libpcp/src/pdu.c: fix day zero botch with EINTR
I knew fixing qa/1123 so it was no longer _notrun everywhere was a
risk, and sure enough on some QA Farm hosts running pmchart with
-c "some shell script" -CC was tripping a PM_ERR_IPC in multiple
(but not deterministic) places.

It took me DAYS to get to the bottom of this (see the next libpcp
commit for the gory details), but the root cause was that in
pduread() we were not handing interrupted read() and select() (via
__pmSocketReady()) calls correctly.

And of course pmchart is one of the few PCP apps that might catch a
timer or signal interrupt while doing PMAPI calls, especially in the
metadata operations (need -CC to pmchart, not -C to trip the bug).

Since as far as PDU exchange in libpcp is concerned nothing has
really happened when errno is set to EINTR after read() or select(),
return -1, it is safe to try again, and again if necessary.

The old code resulted in the synchronous PDU exchange between a PMAPI
client (pmchart) and pmcd being broken which triggers a cascade of
PM_ERR_IPC errors at some random point in the future.