libpcp: extra diagnostics associated with PM_ERR_IPC
The EINTR saga with qa/1123 and pmchart highlighted the fact that
when PM_ERR_IPC occurs we hav vanishingly little information to go on.

This is made worse by the fact that PM_ERR_IPC errors are often a
result of timing windows, lock botches in the presence of multiple
threads, DoS attack and bad coding, ... and given the synchronous
nature of the PDU exchanges with pmcd below the PMAPI there is a risk
of delayed or cascaded reporting of the error above the PMAPI.

It turns out that we set PM_ERR_IPC in 170+ (!!!) places within libpcp,
mostly in the PDU send-and-receive pairs (where receiving the wrong
PDU type is the trigger) and the lower-level PDU decode routines for an
incoming PDU (where bad PDU parameter values and data overrun guards
are the triggers).

In the vast majority of these places there was no diagnostic code to
explain where the PM_ERR_IPC is first detected, nor the reason.

This commit fixes that, and now -Dpdu will enable diagnostics on
*every* libpcp code path that results in PM_ERR_IPC.