pmconfig: harden parsing of invalid configuration settings
A mistaken configuration setting (instances = no) releaved an issue
with parsing invalid or unexpected configurations. Add an obvious
quick fix which is good for now but the method might need to be
fully reviewed later and some QA added to better define and handle
how to deal with invalid or mistaken configuration entries.