pmrep conf: correct collectl cpu metrics to match current versions
collectl docs state about "CPU Data, collectl -sc" the following:

cpu
Percent of time the cpu was busy during the current interval averaged across all
CPUs and is actually the total percentage of time the CPU in one of the following:
system, user, nice, irq, soft-irq and steal.  Note that this does NOT include time
spend in I/O wait.

sys
Percentage of time the cpu was executing in system mode during the current
interval.  This includes all those modes as above except user and nice to to
determine the amount of time spent as a user you need to subtract these from
the total cpu field.

Adjust the corresponding metricset formulas accordingly. No QA impact.