libpcp: rework for some AF routines
We call onalarm() in a signal context, and as such we need to be
"async-signal-safe", and we were not ... specifically calling free()
[very naughty] and gettimeofday() (via pmtimevalNow()) [not quite
so naughty].  Fortunately the free() call only happens in the case of
"single shot" events, which in practice means it is likely that only
pmlogger's "log once" metrics are at risk.

Use a lazy garbage collector to delay free() until it is safe
to call, in __pmAFblock() and __pmAFunblock().  Needs another
global static to keep a list of garbage that needs to be reclaimed.

Change pmtimevalNow() to call clock_gettime() instead of gettimeofday().

All of this was triggered by QA failures on *BSD that involved one of
the following non-deterministic error signatures:
  - malloc/free data corruption of various types
  - pmlogger going into an infinite loop

    modified:   src/libpcp/src/AF.c
    modified:   src/libpcp/src/check-statics
    modified:   src/libpcp/src/tv.c