src/pmdas/linux_proc/proc_pid.c: fix for Operation not permitted" errors for proc.fdinfo.* metrics
For defunct processes, /proc/<pid>/fdinfo exists, contains . and ..
entries but no other.

In refresh_proc_pid_fdinfo(), the intial readdir() loop was counting
the . and .. entries. then falling into the second readdir() loop
that skipped these entries, found no valid entries, did not assign
a value to sts, so we return with sts == -1.

Fixed the readdir() looping logic to be correct.

Change the sts protocol to be 0 unless some error is found.

Added some diagnostics.