qa: assorted tests, take better care of /var/log/pcp/NOTICES
Since this file is now expected to be owned $PCP_USER:$PCP_GROUP
we need t be sure that any QA test that dinks with this file
(a) really needs to dink with it, and (b) puts the file ownership
back the way it was when the test started.

Not doing this was causing some fallout in qa/1483 because of
warnings in syslog about permissions/existence of the NOTICES
file.