libpcp: extra pmns locking
For a LONG time, we've had rare, but repeating, QA failures in the
multithread tests, especially qa/595.

Failure signatures have included:
    - truncated application output
    - malloc corruption aborts
    - segv
    - infinite looping
    - thread synchronization stand off

Finally I trapped one of these failures and found TWO treads in
the same libpcp code call path:
    - pmNewContext()
    - initarchive()
    - __pmFixPMNSHashTable()
    - backlink()
so no wonder badness was hapenning.

We call __pmFixPMNSHashTable() from all over the landscape with a
mixture of locks held and not held and no combination of additional
locking for existing locks is safe ... lock recursion will happen
on at least one of the call paths.  But since __pmFixPMNSHashTable()
does not call any routines that acquire further locks it is safe to
add a new local lock to serialize calls in this routine.