src/pmcd/src/pmcd.c: drop some (recently added) warning messages
The warnings that were reported when -C, -L or -M command line
arguments were used (to change pmcd's limits of various types)
are emitted as command line args are processed and before stderr
is attached to the -l (pmcd.log by default) file.

This leads to chatter that systemctl feels obliged to add to the
journal and which trips several QA changes (now that QA forces -C 512
for pmcd).

But the new pmcd.limits.* metrics provide run-time visibility into
pmcd's limits, so the warnings are not really needed.