libpcp: add new __pmCtlDebug()
Used to save and clear, then restore pmDebugOptions.* flags.

Needed in some diagnostic routines, especially when dumping pmResult-style
objects and -Dpdu really gets in the way with the embedded PMAPI calls
to extract metadata to drive the formatting of the output.

The functionality always was in libpcp, just hidden in util.c, so this
change simply exposes an API for others to access the save/restore
functionality.