common.pmcd.pdu: dodge nc(1) if possible
We have a really bad track record with nc(1) in PCP QA, in part because
there are so many damn versions with different command line options and
different semantics, but also because the socket use does not match
a PMAPI client and pmcd interaction (mostly synchronous, which means
the client needs to get the last response before closing the socket).

This has led to test hangs (the socket never gets closed) or race
conditions (the client ends closes the socket before pmcd's done with
the last PDU xmit).

In an earlier commit for pdu-gadget, I snuck in a "small" change to
make it do the socket I/O directly instead of piping the binary data
to nc.

common.pmcd.pdu now uses this appraoch (no nc) for ascii pdu
specifications.