qa/1900: Eliminate the need for libbpf debuginfo
If the libbpf debuginfo is not installed, valgrind returns a backtrace
of hexadecimal numbers. Those numbers do not match the symbolic
functions names used in qa/1900 valgrind suppression file.  Limiting
the backtrace matching in the valgrind suppress file to just the
syscall function to avoid failures when the libbpf debuginfo is not
installed.