libpcp: PDU decode routines ... wind back some max size checks
We had been using INT_MAX as an upper bound for various checks,
but the max PDU size (__pmSetPDUCeiling(0)) seemed like a better bet.

These checks are mostly to stop DoS memory allocation attacks,
NOT buffer overruns which are subject to a raft of different and
independent checks.

Turns out this change was wrong, as some PDUs (not sent TO pmcd, when
the PDU size limit applies) are longer than 64Kbytes (the default max
PDU size for PDUs received by pmcd) and this was getting caught in QA,
esp qa/513 and qa/593.

Revert to INT_MAX (which is insanely high) until we have a better plan.

    modified:   src/libpcp/src/p_idlist.c
    modified:   src/libpcp/src/p_instance.c
    modified:   src/libpcp/src/p_label.c
    modified:   qa/1560.out
    modified:   qa/pdudata/desc_ids
    modified:   qa/pdudata/instance
    modified:   qa/pdudata/instance_req
    modified:   qa/pdudata/label