libpcp: PDU decode routines ... rework some max size checks
Rather than INT_MAX (which does not really prevent much) or
__pmSetPDUCeiling(0) (which only applies to PDUs received by pmcd) it
is better to craft the guards in terms of the actual PDU size received
... in general terms there is a limit on the maximum number of objects
you can encode in a PDU as defined by the PDU length and the minimum
size of an object.

Should be no functional change here, just tighter tests and more
QA coverage.

    modified:   src/libpcp/src/p_desc.c
    modified:   src/libpcp/src/p_idlist.c
    modified:   src/libpcp/src/p_instance.c
    modified:   src/libpcp/src/p_label.c
    modified:   src/libpcp/src/p_lcontrol.c
    modified:   src/libpcp/src/p_pmns.c
    modified:   src/libpcp/src/p_profile.c

    modified:   qa/src/pdu-gadget.c
    modified:   qa/1560.out

    modified:   qa/pdudata/desc_ids
    modified:   qa/pdudata/descs
    modified:   qa/pdudata/instance
    modified:   qa/pdudata/instance_req
    modified:   qa/pdudata/label
    new file:   qa/pdudata/pmns_names
    new file:   qa/pdudata/profile