qa/507: use -h localhost instead of -h $qahost
In CI, for some containers, the FQDN we get for the local
hostname is a very long name like:
fv-az1984-647.ddddlmzljshefgqgcv3uanxwgh.cx.internal.cloudapp.net
that fails a DNS lookup or maps to a different interface IP addr
and we can't connect to pmcd there.

For the purposes of this test -h localhost works just as well.