Makepkgs: connect $PACKAGE_BUILD to package name for deb builds
At some point in the past this got lost ... although it remains for
rpm packages.

$PACKAGE_BUILD from VERSION.pcp should make its way through to the
name of the packages, as in

    PACKAGE_BUILD=8
    ...
    pcp_6.3.1-8_amd64.deb
              ^

Note this is only for developer builds (or pcp-daily) where VERSION.pcp
is updated to reflect the latest spin number.  For everyone else
PACKAGE_BUILD=1 and there is no change.