src/pminfo/pminfo.c: error path rework
Started off with a memory corruption that was triggered by
qa/admin/upgrade-test when installing new PCP packages on
a system where PCP once lived, leading to a mismatch between
pmcd.conf and the PMNS, and a dynamic metric in the PMNS that
had no associated PMDA known to pmcd.

Fixing that proved a bit more difficult, as we have "batched"
and "unbatched" versions of both name lookup and pmDesc retrieval
that have different error protocols.  Needed to play games with
namelist[] and pmidlist[] to avoid "use-after-free" and "leak"
memory conditions in the presence of various error combinations
and command line arguments.